requirements.txt
black==22.10.0
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Cors==3.0.10
#Flask-JWT==0.3.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.4
psycopg2-binary==2.9.4
PyJWT==2.5.0
PyMySQL==1.0.2
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.42
tomli==2.0.1
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.9.0
